NOISSUE(chore): extract tools to separate files, add typing to tools callbacks
